Updating MCP server to use new Transcribe APIs
