+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@aviaryhq/cloudglue-mcp-server",
   "type": "module",
-  "version": "0.0.7",
+  "version": "0.0.8",
   "main": "index.js",
   "bin": {
     "cloudglue-mcp-server": "build/index.js"
@@ -24,7 +24,7 @@
   "license": "Elastic-2.0",
   "description": "CloudGlue MCP Server",
   "dependencies": {
-    "@aviaryhq/cloudglue-js": "^0.0.11",
+    "@aviaryhq/cloudglue-js": "github:aviaryhq/cloudglue-js#c7d1800c563bc8f51bce1f73b616b1359d144fdd",
     "@modelcontextprotocol/sdk": "^1.9.0",
     "dotenv": "^16.4.7",
     "shx": "^0.4.0",
