+++ src/tools/transcribe-cloudglue-video.ts
@@ -1,0 +1,77 @@
+import { z } from "zod";
+import { CloudGlue } from "@aviaryhq/cloudglue-js";
+import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
+
+export const schema = {
+  url: z
+    .string()
+    .describe(
+      "The fully qualified cloudglue url of the video to describe, e.g. cloudglue://files/file-id",
+    ),
+};
+
+export function registerDescribeCloudglueVideo(
+  server: McpServer,
+  cgClient: CloudGlue,
+) {
+  server.tool(
+    "transcribe_cloudglue_video",
+    "Returns rich multimodal video transcript (detailed description) of a video uploaded to CloudGlue. Don't use this tool if the video is already part of a collection; use get_collection_video_description instead.",
+    schema,
+    async ({ url }) => {
+      const transcribeJob = await cgClient.transcribe.createTranscribe(url, {
+        enable_summary: true,
+        enable_speech: true,
+        enable_scene_text: true,
+        enable_visual_scene_description: true,
+      });
+
+      while (
+        transcribeJob.status === "pending" ||
+        transcribeJob.status === "processing"
+      ) {
+        await new Promise((resolve) => setTimeout(resolve, 5000));
+
+        const updatedJob = await cgClient.transcribe.getTranscribe(
+          transcribeJob.job_id,
+          {
+            response_format: "markdown",
+          },
+        );
+        if (updatedJob.status !== transcribeJob.status) {
+          Object.assign(transcribeJob, updatedJob);
+        }
+      }
+
+      if (transcribeJob.status === "completed") {
+        if (transcribeJob.data?.content) {
+          return {
+            content: [
+              {
+                type: "text",
+                text: transcribeJob.data.content,
+              },
+            ],
+          };
+        }
+        return {
+          content: [
+            {
+              type: "text",
+              text: JSON.stringify(transcribeJob.data),
+            },
+          ],
+        };
+      }
+
+      return {
+        content: [
+          {
+            type: "text",
+            text: "Error: Failed to describe video",
+          },
+        ],
+      };
+    },
+  );
+}
